data_crawling/requirements.txt
requests==2.28.2
beautifulsoup4==4.11.1
lxml==4.9.1
fastapi==0.95.0
uvicorn==0.21.1
schedule==1.2.0 
python-dotenv==1.0.1